Fix for parallelism issue #948
Before we had difficulty with the traversal when determining a parallel
block. The algorithm was broken and would often traverse to the end,
including source nodes. This just rewrites it -- I didn't bother
debugigng because the approach was not great from the start. This does a
simple DFS + uses a nonlocal variable to track the origin (and update
it), so we can return it.

See #948 for context.