Fixes issues in which foreign key IDs were not displayed in the UI
This is a bit ugly -- we're still identifying the underlying cause, but
it turns out that in the SQL orm interaction with django ninja, foreign
keys will not show up as IDs.

If you have foreign_key_id: int in the model it'll work. So adding this
in fixed it.