added `plugins.SchemaValidator`and `experimental.h_databackend` (#925)
The main contribution of this commit is the `SchemaValidator` adapter.
It also introduces "databackends".

It can serve as a template for "runtime validator" that save and load
artifacts to conduct validation. In this case, we produce a schema on disk
then subsequent runs read schema from disk for validation. This approach 
contrasts with @check_output decorators that use statically-defined 
checks for validation. 

Components from the diffing algorithm could be reused for other purposes
such as diffing dataflow and node versions.

"databackends" are a way to defined "abstract types" that can be used with
`isinstance()` and `issubclass()` without actually importing the package. In this
case, it allows us to define different implementations to retrieve the schema
(e.g., pandas, polars, Ibis) without having `h_schema.py` depend on any of them.

We already have a similar mechanism for polymorphism in Hamilton which uses
hamilton.registry to register the different hamilton.plugins.FOO_extensions. The
limitation of this approach is that it leads to decentralized definitions where we
we would have to define how to get a schema in each file whereas databackends
allows us to define them all in one file for easier code management.

---------

Co-authored-by: zilto <tjean@DESKTOP-V6JDCS2>