Example showing inline data saver & loaders
This is a proof of concept.

What needs to be actually done:

1. ideally we expand/wrap the function with the dataloader type appropriately,
to mirror the current process (I think that's what we want).

I made them classes to make it easy to add from_X functions to create
the metadata. Otherwise I don't type the metadata dictionaries --
so maybe we should do that / provide a way to push people to putting
standard things in it.

Otherwise I think this is more ergonomic for most people getting
started.