Allows static files to be served in non-debug mode
There is a simple flag that allows it. While we really should be not
using the django server, this is OK for now. This is necessary as django
does not, by default, serve static files in non-debug mode. We want that
to happen so someone can run this. We will be providing a more
production-blessed path soon, but this unblocks our users.