Adds authentication guard + redirects to login in local mode
This is generally a no-op, but it is nice to have for the local
authenticated version. A bit of a hack, but this is really just a quick
way to get something up and running. In the medium term we will be
changing the way we do auth to be more unified.