Adds spark input type validator
Spark chosees between the "connect" and the "classic" sessions -- this
makes the type-checker in hamilton unhappy because spark has not set the
proper subclass relationships.

Here is why it is necessary
    - https://community.databricks.com/t5/data-engineering/pyspark-sql-connect-dataframe-dataframe-vs-pyspark-sql-dataframe/td-p/71055
    - https://issues.apache.org/jira/browse/SPARK-47909

This provides an adapter + an instance of this adapter for easy use.

This also bumps to spark >=3.0, which helps the imports work for tests.