Sketches Schema UI view
This is hacky UI code. It does not handle "run ids".

1. Adds to pandas & pyspark schema capture using `h_schema.py`.
2. Adds SchemaView and related types to try to show a table...

So it's either we've set something up with too many things, or this
is just a factor of using typescript... Basically I dont like how many
things I had to add to do this..

Otherwise it's non-obvious to me what the pattern should be to
handle comparing runs -- this UI change for the schema table
doesn't take that into account -- it should probably use Generic Table
but I couldn't wrap my head around it.

Otherwise open question whether we reuse the h_schema pyarrow
stuff, or just roll our own again...