Refactors capturing of data to be more flexible
We now have three categories:
1. Result summaries -- we can output one and if not it will be
   unsupported in the UI
2. Schema -- we can output one or none
3. Additional -- as many as we want

We use a single dispatch function for each, and it makes the code a lot
cleaner. We no longer put stuff in lists unless its an additional
result. Note they can also supply their own name, if not, we will
generate a unique attribute name.

In the long term we'll likely want to stop using single dispatch as we
need to register multiple with roles (and single dispatch requires one
function per role). For now this is clean enough and easy to work with,
however.