Adds string diff view
This just treats all primitives as strings, which can show the view.
This works with strings and will actually just look fine otherwise. It
uses the ReactDiffViewer component which is simple, and we use elsewhere
in the app.

We can probably tune this a bit (the interaction is a little jumpy), but
for now this is OK.