Plugin/mlflow custom flavor - fixed small bug (#1068)
Thanks for adding the custom flavor option to mlflow. I tested it and only the flavor specific kwargs somehow got mixed up and tried to be called somewhere else which lead to a fatal error.

Changes
renamed class attribute to avoid conflicting calls

How I tested this
added unit test that doubles as minimal example for reproducing the error

Notes
only tested the DataSaver (due to internal legacy structure, DataLoader is not utilised/hard to implement)

* working mlflow plugin need to add kwargs for custom flavor

* flavor kwargs used outside of mlflow saver/loader class

* updated class description

* flavor kwargs used outside of mlflow saver/loader class

* removed incidental files

---------

Co-authored-by: Jernej Frank <jernej.frank@oxehealth.com>