Implements #710 optional value exposure (#1035)
* Implements #710 optional value exposure

This fixes #710.

This hasn't been extensively tested, but this
seems to work and pass things through as
expected. There is a JSON serialization
caveat, however we expect default
values to be primitives so that should
be fine. We default to not returning optional
values in this case, and users must request it
themselves.