Gets rid of ruff type-checking block hint for first-class imports
(a) I don't really like this -- it should be done judiciously and not
litter the code. Circular imports are rare and they should be first
avoided then dealt with in this manner.
(b) It is particularly bad for the Hamilton codebase as types, etc...
are often defined inline and required at runtime.