feat: First-class caching (#1104)
Caching is now a core Hamilton feature. You can add caching
by using `Builder().with_cache()` and the cache is accessible
via the `Driver.cache` attribute.

The feature is implemented using lifecycle hooks, but this could
change in Hamilton 2.0.

This also introduces the concept of `MetadataStore` and `ResultStore`
which could be further reused with the Hamilton UI or in 
specific deployment scenarios.

This also includes a bug fix for the `@dataloader` decorator which
had accidentally flipped two tags compared to `@load_from` and `from_.`.
The fix includes changes to `@dataloader`, a related test, and to the
visualization function in `graph.create_graphviz_graph`.

The reference documentation page `caching-logic.rst` includes many
TODOs, quirks, limitations, and roadmap items. Eventually, we will need
to deprecate the other caching mechanisms.

---------

Co-authored-by: zilto <tjean@DESKTOP-V6JDCS2>