feat: in-memory caching (#1207)
Added in-memory caching. Stores exist in the Python process, but are not 
persisted to disk after each node execution. The entire stored can be persisted
or loaded via new methods.  

Slight refactor to the `HamiltonCacheAdapter` means that implementations
supporting multiprocessing need to implement a `__getstate__` method.
Note that for `Python >= 3.11` all objects have a default implementation of
`__getstate__` but it's not the case for earlier versions.

---------

Co-authored-by: zilto <tjean@DESKTOP-V6JDCS2>