Add with_columns for pandas
Exteded the with_columns to be used on pandas dataframe.

Pandas does not have a native implementation of with_columns. This
builds one by using the existing extract_columns decorator to create
node(s) for each specified column(s) in the dataframe, then a subdag
gets build the usual way and at the end selected end nodes get appended
to a copy of the original dataframe using pandas.assing -- which creates
a shallow copy of the original dataframe. In case columns with the same name
are selected as end nodes it overrides the existing columns.