Fixes #1240 (#1241)
* Fixes #1240

The cache store assumed that every persister took a `path` argument. That is
not the case because the savers / loaders wrap external APIs and we decided
to not try to create our own abstraction layer around them, and instead mirror them.

E.g. polars takes `file`, but pandas takes `path`.

This means future changes could need to change things here.

* Adds tests

To catch case with `file` and without `path` or `file`.