Adds modular subdag example and configuration object for subdag config  (#1251)
* Enables subdag to handle `configuration` and `value` for config values

To be consistent and to allow people to remap external values we need
this functionality.

E.g. subdag has a config key "foo", but we need two different values
because we are reusing it in two subdags. So we can
remap the values from external config:

Global config: {"foo1": "foo_v1", "foo2": "foo_v2"}
Subdag1 config: {"foo": configuration("foo1")}
Subdag2 config: {"foo": configuration{"foo2")}

This also handles if someone wants to use value().

Otherwise to preserve backwards compatibility we need
to do this in a specific order and only override config values
appropriately.

Squashed commits:

* Adds modular model pipeline example

This shows how you can construct a pipeline from components
and then use subdag to parameterize it for reuse.

* Adds more notes  to module notebook example

* Adds configuration object and refactors code

Named it configuration to not clash with config decorator.

* Adds more tests

* Updates modular example image